+++ includes/ecommerce.php
@@ -109,7 +109,7 @@ function sendbeam_ecommerce_active( $event ) {
  */
 function sendbeam_handle_save_ecommerce() {
 	if ( ! current_user_can( 'manage_options' ) ) {
-		wp_die( esc_html__( 'You do not have permission to do that.', 'sendbeam' ) );
+		wp_die( esc_html__( 'You do not have permission to do that.', 'sendbeam' ), '', array( 'response' => 403 ) );
 	}
 	check_admin_referer( 'sendbeam_save_ecommerce' );