+++ includes/mail-log.php
@@ -505,7 +505,7 @@ function sendbeam_mail_log_unschedule_prune() {
  */
 function sendbeam_handle_clear_log() {
 	if ( ! current_user_can( 'manage_options' ) ) {
-		wp_die( esc_html__( 'You do not have permission to do that.', 'sendbeam' ) );
+		wp_die( esc_html__( 'You do not have permission to do that.', 'sendbeam' ), '', array( 'response' => 403 ) );
 	}
 	check_admin_referer( 'sendbeam_clear_log' );