+++ includes/mail.php
@@ -736,7 +736,7 @@ function sendbeam_mail_error_bundle( $error, $status = 0 ) {
  */
 function sendbeam_handle_test_mail() {
 	if ( ! current_user_can( 'manage_options' ) ) {
-		wp_die( esc_html__( 'You do not have permission to do that.', 'sendbeam' ) );
+		wp_die( esc_html__( 'You do not have permission to do that.', 'sendbeam' ), '', array( 'response' => 403 ) );
 	}
 	check_admin_referer( 'sendbeam_test_mail' );